requirements.txt
Flask==3.0.3
python-docx==1.1.2
docx2txt==0.8
pdfplumber==0.11.4
